requirements.txt
sentence-transformers
datasets
psycopg2-binary
sqlalchemy
ollama
langchain
